Remove AgentManager and related chat functionality to streamline agent management
