+++ agent-mimir/src/communication/multi-agent.ts
